Reset All: converted it to a component.
To make App.vue as "clean" as possible.
