+++ src/renderer/src/App.vue
@@ -1,30 +1,4 @@
-<script setup>
-import { useConfirm } from 'primevue/useconfirm'
-
-const confirm = useConfirm()
-
-if (window.electronAPI !== undefined) {
-  window.electronAPI.onResetAll(() => {
-    confirm.require({
-      header: 'OpenCOR',
-      icon: 'pi pi-question-circle',
-      message: 'You are about to reset all of your settings. Do you want to proceed?',
-      acceptProps: {
-        severity: 'danger'
-      },
-      rejectProps: {
-        severity: 'secondary',
-        outlined: true
-      },
-      accept: () => {
-        window.electronAPI.resetAll()
-      }
-    })
-  })
-}
-</script>
-
 <template>
   <Background />
-  <ConfirmDialog style="max-width: 75vw" />
+  <ResetAllDialog />
 </template>
